requirements.txt

aiogram==2.25.1
pandas
openai
python-dotenv
